% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{SLMRtoSAS}
\alias{SLMRtoSAS}
\title{Convert R SLM code to SAS}
\usage{
SLMRtoSAS(lm)
}
\arguments{
\item{lm}{a simple linear model}
}
\value{
'sascode.txt' the simple linear model in SAS code
}
\description{
Convert the simple linear model R code to SAS code.
}
\details{
Extract the factor and numeric variables from a simple linear model R code
and re-built and export the SAS code for this simple linear model in txt form.
The function can deal with factor and numeric variables as well as interaction terms
(up to 3 interaction terms).
}
\examples{
lm1 <- lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width + Species,data = iris)
SLMRtoSAS(lm1)
lm2 <- lm(Sepal.Length ~ . ,data=iris)
SLMRtoSAS(lm2)
lm3 <- lm(Sepal.Length ~ Sepal.Width * Petal.Length * Petal.Width + Species,data = iris)
SLMRtoSAS(lm3)
lm4 <- lm(Sepal.Length ~ Sepal.Width * Petal.Length + Petal.Width * Species,data = iris)
SLMRtoSAS(lm4)

}
\author{
Haoting Han and Zhengya Gao
}
